feat: move saveAppActions in fixture instead of index test file
