+++ src/services/item/plugins/action/base-analytics.test.ts
@@ -6,7 +6,7 @@ import { ChatMessageRepository } from '../../../chat/repository';
 import { Member } from '../../../member/entities/member';
 import { saveMembers } from '../../../member/test/fixtures/members';
 import { Item } from '../../entities/Item';
-import { saveAppActions } from '../app/appAction/test/index.test';
+import { saveAppActions } from '../app/appAction/test/fixtures';
 import { saveAppData } from '../app/appData/test/index.test';
 import { saveAppSettings } from '../app/appSetting/test/index.test';
 import { BaseAnalytics } from './base-analytics';
