+++ src/services/item/plugins/app/appAction/test/fixtures.ts
@@ -1,5 +1,9 @@
 import { v4 } from 'uuid';
 
+import { Member } from '../../../../../member/entities/member';
+import { Item } from '../../../../entities/Item';
+import { AppActionRepository } from '../repository';
+
 export const MOCK_APP_ORIGIN = 'https://app.localhost:3000';
 
 export const MOCK_CONTEXT = {
@@ -24,3 +28,11 @@ export const MOCK_CONTEXT = {
 export const MOCK_APPS = [
   { id: v4(), name: 'some-name', url: 'some-url', description: 'description', extra: {} },
 ];
+
+export const saveAppActions = async ({ item, member }: { item: Item; member?: Member }) => {
+  const defaultData = { type: 'some-type', data: { some: 'data' } };
+  const s1 = await AppActionRepository.save({ item, member, ...defaultData });
+  const s2 = await AppActionRepository.save({ item, member, ...defaultData });
+  const s3 = await AppActionRepository.save({ item, member, ...defaultData });
+  return [s1, s2, s3];
+};
